refactor: Move to service folder
