+++ lib/src/internal/services/event_emitter.dart
